stitching

+++ QuadTree.cpp
@@ -29,8 +29,8 @@ using namespace std;
  */
 std::vector<glm::vec3> QuadTree::normals;
 std::vector<GLfloat> QuadTree::noises;
-std::vector<GLushort> QuadTree::indices;
-std::vector<GLushort> QuadTree::normalIndices;
+std::vector<GLuint> QuadTree::indices;
+std::vector<GLuint> QuadTree::normalIndices;
 std::vector<QuadTree*> QuadTree::quadTreeList;
 Verts QuadTree::verts;
